requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
# numpy==1.24.4
# opencv-python==4.8.1.78
Jinja2==3.1.2
# itsdangerous==2.1.2
# click==8.1.7
google-generativeai>=0.3.0
folium==0.14.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3 